refactor: faliure code를 failure 패키지로 묶기
