+++ src/main/java/sopt/makers/jwt/verifier/application/JwtVerifier.java
@@ -14,7 +14,7 @@ import org.springframework.stereotype.Component;
 
 import java.security.PublicKey;
 import java.text.ParseException;
-import sopt.makers.jwt.verifier.code.JwtFailure;
+import sopt.makers.jwt.verifier.code.failure.JwtFailure;
 import sopt.makers.jwt.verifier.exception.JwtException;
 import sopt.makers.jwt.verifier.exception.JwkException;
 import sopt.makers.jwt.verifier.infrastructure.JwkProvider;
