+++ src/main/java/sopt/makers/jwt/verifier/code/failure/JwkFailure.java
@@ -1,4 +1,4 @@
-package sopt.makers.jwt.verifier.code.jwk;
+package sopt.makers.jwt.verifier.code.failure;
 
 import static lombok.AccessLevel.PRIVATE;
