+++ src/main/java/sopt/makers/jwt/verifier/exception/JwkException.java
@@ -1,6 +1,6 @@
 package sopt.makers.jwt.verifier.exception;
 
-import sopt.makers.jwt.verifier.code.jwk.JwkFailure;
+import sopt.makers.jwt.verifier.code.failure.JwkFailure;
 
 public class JwkException extends BaseException {
     public JwkException(JwkFailure failure) {
