+++ src/main/java/sopt/makers/jwt/verifier/infrastructure/JwkProvider.java
@@ -13,7 +13,7 @@ import java.text.ParseException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import sopt.makers.jwt.verifier.code.jwk.JwkFailure;
+import sopt.makers.jwt.verifier.code.failure.JwkFailure;
 import sopt.makers.jwt.verifier.exception.JwkException;
 
 import java.security.PublicKey;
